requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
click==8.1.8
colorama==0.4.6
ecdsa==0.19.1
executing==2.2.0
fastapi==0.115.11
h11==0.14.0
icecream==2.1.4
idna==3.10
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-jose==3.4.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
